update file permissions

+++ libnczarr/zs3sdk.cpp
@@ -18,10 +18,12 @@
 
 static Aws::SDKOptions zs3options;
 
+/* Forward */
 static Aws::S3::Model::BucketLocationConstraint s3findregion(const char* name);
 static int s3objectsinfo(Aws::Vector<Aws::S3::Model::Object> list, const char* prefix, char*** keysp, size64_t** lenp);
 static void freestringenvv(char** ss);
-
+static int getbucket(const char* pathkey, char** bucketp, const char** keyp);
+    
 #ifndef nullfree
 #define nullfree(x) {if(x) {free(x);}}
 #endif
@@ -175,15 +177,20 @@ NCZ_s3sdkbucketdelete(void* s3client0, void* config0, const char* region, const 
 /* Object API */
 
 int
-NCZ_s3sdkinfo(void* s3client0, const char* bucket, const char* key, size64_t* lenp, char** errmsgp)
+NCZ_s3sdkinfo(void* s3client0, const char* pathkey, size64_t* lenp, char** errmsgp)
 {
     int stat = NC_NOERR;
     Aws::S3::S3Client* s3client = (Aws::S3::S3Client*)s3client0;
     Aws::S3::Model::ListObjectsV2Request objects_request;
     size64_t* lengths = NULL;
     size_t nkeys;
+    char* bucket = NULL;
+    const char* key = NULL;
 
-    if(*key == '/') key++;
+    if(*key != '/') return NC_EINTERNAL;
+    /* extract the bucket prefix */
+    if((stat = getbucket(pathkey,&bucket,&key))) return stat;
+
     if(errmsgp) *errmsgp = NULL;
     objects_request.SetBucket(bucket);
     objects_request.SetPrefix(key);
@@ -193,28 +200,33 @@ NCZ_s3sdkinfo(void* s3client0, const char* bucket, const char* key, size64_t* le
         Aws::Vector<Aws::S3::Model::Object> object_list =
             list_outcome.GetResult().GetContents();
 	nkeys = (size_t)object_list.size();
-	if(nkeys != 1) {stat = NC_EACCESS; goto done;}	
-	if((stat = s3objectsinfo(object_list,"/",NULL,&lengths))) goto done;
-	if(lenp) *lenp = lengths[0];	
+	if(nkeys != 1) {stat = NC_EACCESS;}
+	else {
+	  if(!(stat = s3objectsinfo(object_list,pathkey,NULL,&lengths))) {
+	      if(lenp) *lenp = lengths[0];
+	  }
+	}
     } else {
 	if(errmsgp) *errmsgp = makeerrmsg(list_outcome.GetError(),key);
         stat = NC_ES3;
-        goto done;
     }
-done:
     nullfree(lengths);
     return (stat);
 }
 
 int
-NCZ_s3sdkread(void* s3client0, const char* bucket, const char* key, size64_t start, size64_t count, void* content, char** errmsgp)
+NCZ_s3sdkread(void* s3client0, const char* pathkey, size64_t start, size64_t count, void* content, char** errmsgp)
 {
     int stat = NC_NOERR;
     Aws::S3::S3Client* s3client = (Aws::S3::S3Client*)s3client0;
     Aws::S3::Model::GetObjectRequest object_request;
     char range[1024];
+    char* bucket = NULL;
+    const char* key = NULL;
 
-    if(*key == '/') key++;    
+    if(*key != '/') return NC_EINTERNAL;
+    if(!(stat = getbucket(pathkey,&bucket,&key))) return stat;
+    
     object_request.SetBucket(bucket);
     object_request.SetKey(key);
     snprintf(range,sizeof(range),"bytes=%llu-%llu",start,(start+count)-1);
@@ -222,12 +234,11 @@ NCZ_s3sdkread(void* s3client0, const char* bucket, const char* key, size64_t sta
     auto get_object_result = s3client->GetObject(object_request);
     if(!get_object_result.IsSuccess()) {
 	if(errmsgp) *errmsgp = makeerrmsg(get_object_result.GetError(),key);
-        stat = NC_ES3;
-        goto done;
+	stat = NC_ES3;
     } else {
 	/* Get the whole result */
-        Aws::IOStream &result = get_object_result.GetResultWithOwnership().GetBody();
-        std::string str((std::istreambuf_iterator<char>(result)),std::istreambuf_iterator<char>());
+	Aws::IOStream &result = get_object_result.GetResultWithOwnership().GetBody();
+	std::string str((std::istreambuf_iterator<char>(result)),std::istreambuf_iterator<char>());
 	/* Verify actual result size */
 	size_t slen = str.size();
 	if(slen > count) return NC_ES3;
@@ -235,18 +246,22 @@ NCZ_s3sdkread(void* s3client0, const char* bucket, const char* key, size64_t sta
 	if(content)
 	    memcpy(content,s,slen);
     }
-done:
     return (stat);
 }
 
 int
-NCZ_s3sdkreadobject(void* s3client0, const char* bucket, const char* key, size64_t* sizep, void** contentp, char** errmsgp)
+NCZ_s3sdkreadobject(void* s3client0, const char* pathkey, size64_t* sizep, void** contentp, char** errmsgp)
 {
     int stat = NC_NOERR;
     Aws::S3::S3Client* s3client = (Aws::S3::S3Client*)s3client0;
     Aws::S3::Model::GetObjectRequest object_request;
     size64_t size, red;
     void* content = NULL;
+    char* bucket = NULL;
+    const char* key = NULL;
+
+    if(*key != '/') return NC_EINTERNAL;
+    if(!(stat = getbucket(pathkey,&bucket,&key))) return stat;
 
     if(errmsgp) *errmsgp = NULL;
     object_request.SetBucket(bucket);
@@ -255,7 +270,6 @@ NCZ_s3sdkreadobject(void* s3client0, const char* bucket, const char* key, size64
     if(!get_result.IsSuccess()) {
 	if(errmsgp) *errmsgp = makeerrmsg(get_result.GetError(),key);
         stat = NC_ES3;
-        goto done;
     } else {
 	/* Get the size */
 	size = (size64_t)get_result.GetResult().GetContentLength();	
@@ -271,13 +285,16 @@ NCZ_s3sdkreadobject(void* s3client0, const char* bucket, const char* key, size64
         red = result.rdbuf()->pubseekoff(0,std::ios::ios_base::end);
         result.rdbuf()->pubseekoff(0,std::ios::ios_base::beg); /* reset for reading */
 	if((content = malloc(red))==NULL)
-	    {stat = NC_ENOMEM; goto done;}
-	result.rdbuf()->sgetn((char*)content,red);
+	    stat = NC_ENOMEM;
+	else
+   	    result.rdbuf()->sgetn((char*)content,red);
 #endif
-	/* Verify actual result size */
-	if(red != size) {stat = NC_ES3; goto done;}
-	if(sizep) *sizep = red;
-	if(contentp) {*contentp = content; content = NULL;}
+	if(!stat) {
+  	    /* Verify actual result size */
+	    if(red != size) {stat = NC_ES3; goto done;}
+	    if(sizep) *sizep = red;
+	    if(contentp) {*contentp = content; content = NULL;}
+	}
     }
 done:
     nullfree(content);
@@ -289,12 +306,17 @@ For S3, I can see no way to do a byterange write;
 so we are effectively writing the whole object
 */
 int
-NCZ_s3sdkwriteobject(void* s3client0, const char* bucket, const char* key,  size64_t count, const void* content, char** errmsgp)
+NCZ_s3sdkwriteobject(void* s3client0, const char* pathkey,  size64_t count, const void* content, char** errmsgp)
 {
     int stat = NC_NOERR;
     Aws::S3::S3Client* s3client = (Aws::S3::S3Client*)s3client0;
     Aws::S3::Model::PutObjectRequest object_request;
+    char* bucket = NULL;
+    const char* key = NULL;
 
+    if(*key != '/') return NC_EINTERNAL;
+    if(!(stat = getbucket(pathkey,&bucket,&key))) return stat;
+    
     if(errmsgp) *errmsgp = NULL;
     object_request.SetBucket(bucket);
     object_request.SetKey(key);
@@ -305,12 +327,9 @@ NCZ_s3sdkwriteobject(void* s3client0, const char* bucket, const char* key,  size
     object_request.SetBody(data);
     auto object_result = s3client->PutObject(object_request);
     if(!object_result.IsSuccess()) {
-	if(errmsgp) *errmsgp = makeerrmsg(object_result.GetError(),key);
+        if(errmsgp) *errmsgp = makeerrmsg(object_result.GetError(),key);
         stat = NC_ES3;
-        goto done;
     }
-
-done:
     return (stat);
 }
 
@@ -331,54 +350,59 @@ In theory, the returned list should be sorted in lexical order,
 but it possible that it is not.
 */
 int
-NCZ_s3sdkgetkeys(void* s3client0, const char* bucket, const char* subtree, const char* prefix, size_t* nkeysp, char*** keysp, char** errmsgp)
+NCZ_s3sdkgetkeys(void* s3client0, const char* prefix, size_t* nkeysp, char*** keysp, char** errmsgp)
 {
     int stat = NC_NOERR;
     Aws::S3::S3Client* s3client = (Aws::S3::S3Client*)s3client0;
     Aws::S3::Model::ListObjectsV2Request objects_request;
     size_t nkeys = 0;
+    char* bucket = NULL;
+    const char* rootkey = NULL;
+
+    if(*prefix != '/') return NC_EINTERNAL;
+    if(!(stat = getbucket(prefix,&bucket,&rootkey))) return stat;
 
     if(errmsgp) *errmsgp = NULL;
     objects_request.SetBucket(bucket);
-    objects_request.SetPrefix(subtree);
+    objects_request.SetPrefix(rootkey);
     auto objects_outcome = s3client->ListObjectsV2(objects_request);
     if(objects_outcome.IsSuccess()) {
         Aws::Vector<Aws::S3::Model::Object> object_list =
             objects_outcome.GetResult().GetContents();
-	nkeys = (size_t)object_list.size();
-	if(nkeysp) *nkeysp = nkeys;
-	if((stat = s3objectsinfo(object_list,prefix,keysp,NULL))) goto done;
+        nkeys = (size_t)object_list.size();
+        if(nkeysp) *nkeysp = nkeys;
+        stat = s3objectsinfo(object_list,prefix,keysp,NULL);
     } else {
-	if(errmsgp) *errmsgp = makeerrmsg(objects_outcome.GetError());
+        if(errmsgp) *errmsgp = makeerrmsg(objects_outcome.GetError());
         stat = NC_ES3;
-        goto done;
     }
-done:
     return stat;
 }
 
 int
-NCZ_s3sdkdeletekey(void* s3client0, const char* bucket, const char* key, char** errmsgp)
+NCZ_s3sdkdeletekey(void* s3client0, const char* pathkey, char** errmsgp)
 {
     int stat = NC_NOERR;
     Aws::S3::S3Client* s3client = (Aws::S3::S3Client*)s3client0;
     Aws::S3::Model::DeleteObjectRequest delete_request;
 
 #ifdef NOOP
-    /* Delete this key object */
-    delete_request.SetBucket(bucket);
-    delete_request.SetKey(key);
-    auto delete_result = s3client->DeleteObject(delete_request);
-    if(!delete_result.IsSuccess()) {
-	if(errmsgp) *errmsgp = makeerrmsg(delete_result.GetError(),key);
-        stat = NC_ES3;
-	goto done;
+    char* bucket = NULL;
+    const char* key = NULL;
+    if(!(stat = getbucket(pathkey,&bucket,&key))) {
+        /* Delete this key object */
+        delete_request.SetBucket(bucket);
+        delete_request.SetKey(key);
+        auto delete_result = s3client->DeleteObject(delete_request);
+        if(!delete_result.IsSuccess()) {
+            if(errmsgp) *errmsgp = makeerrmsg(delete_result.GetError(),key);
+            stat = NC_ES3;
+	}
     }
 #else
     fprintf(stderr,"delete object: %s.%s\n",bucket,key); fflush(stderr);
 #endif
 
-done:
     return stat;
 }
 
@@ -398,36 +422,44 @@ s3objectsinfo(Aws::Vector<Aws::S3::Model::Object> list, const char* prefix, char
     nkeys = list.size();
     if(keysp) {
         if((keys=(char**)calloc(sizeof(char*),(nkeys+1)))==NULL)
-	    {stat = NC_ENOMEM; goto done;}
+            stat = NC_ENOMEM;
     }
-    if(lenp) {
-        if((lengths=(size64_t*)calloc(sizeof(size64_t),(nkeys)))==NULL)
-	    {stat = NC_ENOMEM; goto done;}
+    if(!stat) {
+        if(lenp) {
+            if((lengths=(size64_t*)calloc(sizeof(size64_t),(nkeys)))==NULL)
+                stat = NC_ENOMEM;
+        }
     }
-    i = 0;
-    for (auto const &s3_object : list) {
-	char* cstr = NULL;
-	if(keysp) {
-            auto s = s3_object.GetKey();
-	    if(prefix) {
-	        size_t klen = s.length();
-	        cstr = (char*)malloc(klen+prelen+1);
-	        strcpy(cstr,prefix);
-	        strcat(cstr,s.c_str());
-            } else		
-	        cstr = strdup(s.c_str());
-            if(cstr == NULL) 
-	        {stat = NC_ENOMEM; goto done;}
-	    keys[i] = cstr;
-  	    cstr = NULL;
+    if(!stat)  {
+        i = 0;
+        for (auto const &s3_object : list) {
+            char* cstr = NULL;
+            if(keysp) {
+                auto s = s3_object.GetKey();
+                if(prefix) {
+                    size_t klen = s.length();
+                    cstr = (char*)malloc(klen+prelen+1);
+                    strcpy(cstr,prefix);
+                    strcat(cstr,s.c_str());
+                } else              
+                    cstr = strdup(s.c_str());
+                if(cstr == NULL) 
+                    {stat = NC_ENOMEM; break;}
+                keys[i] = cstr;
+                cstr = NULL;
+	    }
+            if(!stat) {
+                if(lenp) lengths[i] = (size64_t)s3_object.GetSize();
+                i++;
+	    }
 	}
-	if(lenp) lengths[i] = (size64_t)s3_object.GetSize();
-	i++;
     }
-    if(keysp) keys[nkeys] = NULL;
-    if(keysp) {*keysp = keys; keys = NULL;}
-    if(lenp) {*lenp = lengths; lengths = NULL;}
-done:
+    if(!stat) {
+        if(keysp) keys[nkeys] = NULL;
+        if(keysp) {*keysp = keys; keys = NULL;}
+        if(lenp) {*lenp = lengths; lengths = NULL;}
+    }
+
     if(keys != NULL) freestringenvv(keys);
     if(lengths != NULL) free(lengths);
     return stat;
@@ -455,4 +487,23 @@ freestringenvv(char** ss)
             nullfree(*p);
         free(ss);
     }
+}
+
+static int
+getbucket(const char* pathkey, char** bucketp, const char** keyp)
+{
+    const char* p;
+    assert(pathkey != NULL && pathkey[0] == '/');
+    p = strchr(pathkey+1,'/'); /* find end of the bucket */
+    assert(p != NULL);
+    if(keyp) *keyp = p+1;
+    if(bucketp) {
+        char* bucket = NULL;
+        ptrdiff_t len = ((p - pathkey) - 1);
+        if((bucket = (char*)malloc(len+1))==NULL) {return NC_ENOMEM;}
+        memcpy(bucket,pathkey+1,len);
+        bucket[len] = '\0';
+        *bucketp = bucket; bucket = NULL;
+    }
+    return NC_NOERR;    
 }
